[Test] 프론트 연동 긴급 테스트
Former-commit-id: 1d88ca518000ae5bbecb85717c19e35044d09d3d

+++ BackEnd/src/main/java/com/orialz/backend/blur/dto/response/FindKeywordGetResponse.java
@@ -13,8 +13,8 @@ import lombok.Setter;
 @Builder
 public class FindKeywordGetResponse {
 	String label;
-	long x;
-	long y;
-	long w;
-	long h;
+	double x;
+	double y;
+	double w;
+	double h;
 }
